init_window: Add a counter for frames in flight
Add a counter for in flight frames and limit to 8 before we throw
them away. 8 is probably more than is sensible for most use but
it allows the user to implement their own, possibly more nuanced
control before we use our sledgehammer.

Also adds a call to retrieve the current count.